requirements.txt
flask_socketio~=5.3.6
flask~=3.0.0
markupsafe~=2.1.3
psutil~=5.9.7
python-dotenv~=1.0.0
torch~=2.1.1+cu121
transformers~=4.35.0
waitress~=2.1.2
werkzeug~=3.0.1
